pyproject.toml
[project]
name = "strands-agents-memory"
version = "0.1.0"
description = "Movie recommendation agent showcasing Strands memory capabilities"
requires-python = ">=3.10"
dependencies = [
    "strands-agents>=1.0.0",
    "strands-agents-tools>=0.2.0",
    "mem0ai",
    "opensearch-py>=3.0.0",
    "qdrant-client>=1.15.1",
    "sentence-transformers>=5.0.0",
    "faiss-cpu>=1.11.0.post1",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["main", "movie_database"]

[tool.setuptools.packages.find]
where = ["."]
include = ["tools*"]
